Brat shiny HTML app.

Call brainy, the brat shiny app to annotate multiple documents.

brainy(doc_dir, coll_data, theme = "paper")

Arguments

doc_dir
Diretory with plaint text documents (.txt), annotation files (.ann) and rds files with Annotation objects stored using saveRDS().

coll_data
Configuration of annotation scheme.

theme
The theme passed into shinythemes.

Details

The function takes a document directory (doc_dir) as input. Plain text files present in the directory (.txt file extension) will be available as documents to be annotated.

Annotations for these documents will be stored in the same directory using the brat standoff annotation format (.ann file extension). The name of the annotation file will be identical with text document, just using the .ann file extension. Annotation files that are present before annotating documents will be parsed and annotations are shown.

If the document has already been tokenized, use the Annotation class of the NLP package to infuse this information for the annotation exercise. In this case, when creating a new annotation, the span of text highlighted will be checked against the tokenization and character offsets will be mapped on the existing tokenisation. This can also be useful for speeding up annotation, because annotators need not highlight the exact beginning and end of tokens to create an annotation that looks all right.

Files with an existing token annotation are expected to have the same file name as the document, using an .rds file extension. These rds files are expected to be Annotation class objects that have been stored using saveRDS().

Examples

library(brat)

# Prepare directory with .txt and .ann files: The package includes some 
# sample files - we create a temporary copy

doc_dir_pkg <- system.file(package = "brat", "extdata", "sample_data")
doc_dir <- file.path(tempdir(), "brat")
if (!dir.exists(doc_dir)) dir.create(doc_dir)
if (length(list.files(doc_dir)) > 0L) file.remove(list.files(doc_dir, full.names = TRUE))
for (file in list.files(doc_dir_pkg, full.names = TRUE)){
  file.copy(from = file, to = file.path(doc_dir, basename(file)))
}

# Define collection data

coll_data <- list(
  entity_types = list(
    list(
      type = "Person",
      labels = c("Person", "Per"),
      bgColor = RColorBrewer::brewer.pal(8, "Accent")[1],
      borderColor = "darken",
      arcs = list(list(targets = list("Person"))),
      children = list()
    ),
    list(
      type = "Organization",
      labels = c("Organization", "Org"),
      bgColor = RColorBrewer::brewer.pal(8, "Accent")[2],
      borderColor = "darken",
      arcs = list(list(targets = list("Person"))),
      children = list()
    ),
    list(
      type = "Location",
      labels = c("Location", "Loc"),
      bgColor = RColorBrewer::brewer.pal(8, "Accent")[3],
      borderColor = "darken",
      arcs = list(list(targets = list("Person"))),
      children = list()
    ),
    list(
      type = "Date",
      labels = c("Date", "Date"),
      bgColor = RColorBrewer::brewer.pal(8, "Accent")[4],
      borderColor = "darken",
      arcs = list(list(targets = list("Person"))),
      children = list()
    )
  ),
  relation_types = list(
    list(
      type = "Anaphora",
      labels = c("Anaphora", "Ana"),
      dashArray = "3,3",
      color = "purple",
      args = list(
        list(role = "Anaphor", targets = "Person"),
        list(role = "Entity", targets = "Person")
      )
     ),
     list(
      type = "Birthday",
      labels = c("Birthday", "born"),
      dashArray = "3,3",
      color = "blue",
      args = list(
        list(role = "Anaphor", targets = "Person"),
        list(role = "Entity", targets = "Person")
      )
    )
  )
)

# Run brainy app (but only in interactive mode)

if (interactive()){
  brainy(doc_dir = doc_dir, coll_data = coll_data)
}